+++ test/framework/src/main/java/org/opensearch/test/InternalTestCluster.java
@@ -165,7 +165,6 @@ import static org.opensearch.test.NodeRoles.onlyRole;
 import static org.opensearch.test.NodeRoles.onlyRoles;
 import static org.opensearch.test.NodeRoles.removeRoles;
 import static org.opensearch.test.OpenSearchTestCase.assertBusy;
-import static org.opensearch.test.OpenSearchTestCase.getBaseStreamPort;
 import static org.opensearch.test.OpenSearchTestCase.randomBoolean;
 import static org.opensearch.test.OpenSearchTestCase.randomFrom;
 import static org.hamcrest.Matchers.equalTo;
@@ -758,7 +757,6 @@ public final class InternalTestCluster extends TestCluster {
         final Settings.Builder updatedSettings = Settings.builder();
 
         updatedSettings.put(Environment.PATH_HOME_SETTING.getKey(), baseDir);
-        updatedSettings.put("node.attr.transport.stream.port", getBaseStreamPort() + FLIGHT_PORT_COUNTER.getAndIncrement());
         if (numDataPaths > 1) {
             updatedSettings.putList(
                 Environment.PATH_DATA_SETTING.getKey(),